+++ allennlp/data/iterators/bucket_iterator.py
@@ -6,7 +6,7 @@ from typing import List, Tuple, Iterable, cast, Dict, Deque
 from overrides import overrides
 
 from allennlp.common.util import lazy_groups_of, add_noise_to_dict_values
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 from allennlp.data.instance import Instance
 from allennlp.data.iterators.data_iterator import DataIterator
