+++ allennlp/data/iterators/homogeneous_batch_iterator.py
@@ -3,7 +3,7 @@ import random
 from collections import defaultdict
 
 from allennlp.common.util import lazy_groups_of
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 from allennlp.data.instance import Instance
 from allennlp.data.iterators.data_iterator import DataIterator
