+++ allennlp/data/iterators/pass_through_iterator.py
@@ -4,7 +4,7 @@ import logging
 
 from overrides import overrides
 
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 from allennlp.data.instance import Instance
 from allennlp.data.iterators.data_iterator import DataIterator, TensorDict
