+++ allennlp/data/iterators/same_language_iterator.py
@@ -7,7 +7,7 @@ from allennlp.common.util import lazy_groups_of
 from allennlp.data.instance import Instance
 from allennlp.data.iterators.data_iterator import DataIterator
 from allennlp.data.iterators.bucket_iterator import BucketIterator
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 
 logger = logging.getLogger(__name__)
