+++ allennlp/models/model.py
@@ -14,7 +14,7 @@ from allennlp.common.checks import ConfigurationError
 from allennlp.common.params import Params
 from allennlp.common.registrable import Registrable
 from allennlp.data import Instance, Vocabulary
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 from allennlp.nn import util
 from allennlp.nn.regularizers import RegularizerApplicator
