+++ allennlp/tests/modules/token_embedders/elmo_token_embedder_test.py
@@ -9,7 +9,7 @@ import torch
 from allennlp.commands.train import train_model
 from allennlp.common import Params
 from allennlp.common.testing import ModelTestCase
-from allennlp.data.dataset import Batch
+from allennlp.data.batch import Batch
 from allennlp.modules.token_embedders import ElmoTokenEmbedder
